fix: authenticate  to frienship routers

+++ src/main/java/com/concord/concordapi/friendship/service/FriendshipService.java
@@ -1,10 +1,12 @@
 package com.concord.concordapi.friendship.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authorization.AuthorizationDeniedException;
 import org.springframework.stereotype.Service;
 import com.concord.concordapi.friendship.entity.Friendship;
 import com.concord.concordapi.friendship.repository.FriendshipRepository;
 import com.concord.concordapi.shared.exception.EntityNotFoundException;
+import com.concord.concordapi.auth.service.AuthService;
 import com.concord.concordapi.friendship.dto.FriendshipCreateDTO;
 import com.concord.concordapi.friendship.dto.FriendshipDTO;
 import com.concord.concordapi.friendship.dto.FriendshipPutDTO;
@@ -22,6 +24,8 @@ public class FriendshipService {
     private FriendshipRepository friendshipRepository;
     @Autowired
     private UserRepository userRepository;
+    @Autowired
+    private AuthService authInfoService;
 
     public FriendshipDTO get(Long id) {
         Friendship friendship = friendshipRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Friendship not found"));
@@ -44,7 +48,11 @@ public class FriendshipService {
 
     public FriendshipDTO create(FriendshipCreateDTO friendshipDTO) {
         Friendship friendship = new Friendship();
+        
         User from = userRepository.findById(friendshipDTO.fromId()).orElseThrow(()-> new EntityNotFoundException("User id "+friendshipDTO.fromId()+" not found"));
+        if (!from.getUsername().equals(authInfoService.getAuthenticatedUsername())) {
+            throw new AuthorizationDeniedException("User doesn't match the logged-in user");
+        }
         User to = userRepository.findById(friendshipDTO.toId()).orElseThrow(()-> new EntityNotFoundException("User id "+friendshipDTO.toId()+" not found"));
         friendship.setFrom_user(from);
         friendship.setTo_user(to);
@@ -56,12 +64,20 @@ public class FriendshipService {
     }
 
     public void delete(Long id) {
+        User from = userRepository.findById(id).orElseThrow(()-> new EntityNotFoundException("User id "+id+" not found"));
+        if (!from.getUsername().equals(authInfoService.getAuthenticatedUsername())) {
+            throw new AuthorizationDeniedException("User doesn't match the logged-in user");
+        }
         Friendship friendship = friendshipRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Friendship id "+id+" not found"));
         friendshipRepository.delete(friendship); // Deleta a amizade
     }
 
     public FriendshipDTO update(Long id, FriendshipPutDTO friendshipDTO) {
+        User from = userRepository.findById(id).orElseThrow(()-> new EntityNotFoundException("User id "+id+" not found"));
+        if (!from.getUsername().equals(authInfoService.getAuthenticatedUsername())) {
+            throw new AuthorizationDeniedException("User doesn't match the logged-in user");
+        }
         Friendship friendship = friendshipRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Friendship id "+id+" not found"));
         friendship.setStatus(friendshipDTO.status());
